demo: Move SAH cost computation to main demo
This code really relates to measuring the efficiency characteristics of
an algorithm so it belongs in the main demo file for now. This adds a
lot of code that generally speaking isn't very relevant to the usage of
the library but we should be able to simplify this somewhat.
